polymarket-wallet-alpha-watch/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "polymarket-wallet-alpha-watch"
version = "0.1.0"
description = "Topic-focused Polymarket wallet discovery and PnL export"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "polymarket-wallet-alpha-watch" }]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "pandas>=2.2.0",
  "openpyxl>=3.1.0",
  "typer>=0.12.3",
  "PyYAML>=6.0.2",
  "tenacity>=9.0.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.6.0",
]

[project.scripts]
polyalpha = "polyalpha.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ASYNC"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
